triangulation-3D-visualizer/requirements.txt
depthai==0.2.0.1
pygame==1.9.6
pyopengl==3.1.5